restructure sparse storages module
