+++ lib/segment/src/vector_storage/sparse/mod.rs
@@ -1,0 +1,2 @@
+pub mod mmap_sparse_vector_storage;
+pub mod simple_sparse_vector_storage;
